+++ assigned_bank/__manifest__.py
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright 2019-20 - SHS-AV s.r.l. <https://www.zeroincombenze.it/>
+#
+# Contributions to development, thanks to:
+# * Antonio Maria Vigliotti <antoniomaria.vigliotti@gmail.com>
+#
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+#
+{
+    'name': 'Assigned bank',
+    'summary': 'Assign internal bank to customers or supplier',
+    'version': '10.0.0.1.0',
+    'category': 'Generic Modules/Accounting',
+    'author': 'SHS-AV s.r.l.',
+    'website': 'https://www.zeroincombenze.it/',
+    'depends': ['base',
+                'account'],
+    'data': [
+        'views/partner_view.xml',
+    ],
+    'installable': True,
+}