+++ l10n_it_ddt/__manifest__.py
@@ -5,7 +5,7 @@
 {
     'name': 'DDT',
     'summary': 'Delivery Document to Transfer',
-    'version': '10.0.1.8.2',
+    'version': '10.0.1.8.3',
     'category': 'Localization/Italy',
     'author': 'Odoo Community Association (OCA) and other subjects',
     'website': 'https://www.zeroincombenze.it/servizi-le-imprese/',