+++ l10n_it_lettera_intento/__manifest__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2018-20 - SHS-AV s.r.l. <https://www.zeroincombenze.it/>
+# Copyright 2019-20 - SHS-AV s.r.l. <https://www.zeroincombenze.it/>
 #
 # Contributions to development, thanks to:
 # * Antonio Maria Vigliotti <antoniomaria.vigliotti@gmail.com>
@@ -10,15 +10,17 @@
 {
     'name': 'Lettere di intento',
     'summary': 'Lettere di intento',
-    'version': '10.0.0.1.0',
+    'version': '10.0.0.1.1',
     'category': 'Generic Modules/Accounting',
     'author': 'SHS-AV s.r.l.',
     'website': 'https://www.zeroincombenze.it/',
-    'depends': ['base'],
+    'depends': ['base', 'l10n_it_einvoice_stamp'],
     'data': [
         'security/ir.model.access.csv',
         'data/ir_sequence_data.xml',
         'views/italy_lettera_intento_view.xml',
+        'views/account_fiscal_position_view.xml',
+        'views/config_view.xml',
     ],
     'installable': True,
 }